+++ sei-ibc-go/modules/light-clients/06-solomachine/types/proposal_handle.go
@@ -20,7 +20,6 @@ import (
 func (cs ClientState) CheckSubstituteAndUpdateState(
 	ctx sdk.Context, cdc codec.BinaryCodec, subjectClientStore,
 	_ sdk.KVStore, substituteClient exported.ClientState,
-	_ exported.Height,
 ) (exported.ClientState, error) {
 
 	if !cs.AllowUpdateAfterProposal {